chore(src): reorganize Text Input elements
