+++ src/helix-ui/elements/index.js
@@ -37,7 +37,7 @@ export { HXTabcontentElement } from './hx-tabcontent/index.js';
 export { HXTablistElement } from './hx-tablist/index.js';
 export { HXTabpanelElement } from './hx-tabpanel/index.js';
 export { HXTabsetElement } from './hx-tabset/index.js';
-export { HXTextControlElement } from './HXTextControlElement';
+export { HXTextControlElement } from './hx-text-control/index.js';
 export { HXTextareaControlElement } from './hx-textarea-control/index.js';
 export { HXToastElement } from './hx-toast/index.js';
 export { HXTooltipElement } from './hx-tooltip/index.js';
